+++ lib/src/adaptive_dropdown_search.dart
@@ -24,16 +24,21 @@ class AdaptiveDropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforeChange,
     super.onBeforePopupOpening,
     super.onFocusChange,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.textProps,
   }) : super(
-          popupProps:
-              context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
-                  ? popupProps.cupertinoProps
-                  : popupProps.materialProps,
-          uiMode: UiMode.adaptive,
+          uiMode: context.getUiToApply(UiMode.adaptive),
+          popupProps: context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
+              ? popupProps.cupertinoProps
+              : popupProps.materialProps,
+          groupId: context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
+              ? popupProps.cupertinoProps.autoCompleteProps.groupId
+              : popupProps.materialProps.autoCompleteProps.groupId,
         );
 
   AdaptiveDropdownSearch.multiSelection({
@@ -58,15 +63,21 @@ class AdaptiveDropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforePopupOpening,
     super.onFocusChange,
     super.dropdownBuilder,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.selectedItemsWrapProps,
+    super.textProps,
   }) : super.multiSelection(
-          popupProps:
-              context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
-                  ? popupProps.cupertinoProps
-                  : popupProps.materialProps,
-          uiMode: UiMode.adaptive,
+          uiMode: context.getUiToApply(UiMode.adaptive),
+          popupProps: context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
+              ? popupProps.cupertinoProps
+              : popupProps.materialProps,
+          groupId: context.getUiToApply(UiMode.adaptive) == UiToApply.cupertino
+              ? popupProps.cupertinoProps.autoCompleteProps.groupId
+              : popupProps.materialProps.autoCompleteProps.groupId,
         );
 }