+++ lib/src/cupertino_dropdown_search.dart
@@ -1,4 +1,5 @@
 import 'package:dropdown_search/src/base_dropdown_search.dart';
+import 'package:dropdown_search/src/utils.dart';
 
 import 'widgets/props/cupertino_popup_props.dart';
 
@@ -22,13 +23,16 @@ class CupertinoDropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforeChange,
     super.onBeforePopupOpening,
     super.onFocusChange,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.textProps,
   }) : super(
           popupProps: popupProps,
-          uiMode: UiMode.cupertino,
+          uiMode: UiToApply.cupertino,
           groupId: popupProps.autoCompleteProps.groupId,
         );
 
@@ -53,13 +57,17 @@ class CupertinoDropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforePopupOpening,
     super.onFocusChange,
     super.dropdownBuilder,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.selectedItemsWrapProps,
+    super.textProps,
   }) : super.multiSelection(
           popupProps: popupProps,
-          uiMode: UiMode.cupertino,
+          uiMode: UiToApply.cupertino,
           groupId: popupProps.autoCompleteProps.groupId,
         );
 }