+++ lib/src/dropdown_search.dart
@@ -1,5 +1,6 @@
 import 'package:dropdown_search/src/base_dropdown_search.dart';
 import 'package:dropdown_search/src/popups/props/material_popup_props.dart';
+import 'package:dropdown_search/src/utils.dart';
 
 class DropdownSearch<T> extends BaseDropdownSearch<T> {
   DropdownSearch({
@@ -21,19 +22,21 @@ class DropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforeChange,
     super.onBeforePopupOpening,
     super.onFocusChange,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.textProps,
   }) : super(
           popupProps: popupProps,
-          uiMode: UiMode.material,
+          uiMode: UiToApply.material,
           groupId: popupProps.autoCompleteProps.groupId,
         );
 
   DropdownSearch.multiSelection({
-    MultiSelectionPopupProps<T> popupProps =
-        const MultiSelectionPopupProps.menu(),
+    MultiSelectionPopupProps<T> popupProps = const MultiSelectionPopupProps.menu(),
     super.key,
     super.mode = Mode.form,
     super.autoValidateMode,
@@ -52,13 +55,17 @@ class DropdownSearch<T> extends BaseDropdownSearch<T> {
     super.onBeforePopupOpening,
     super.onFocusChange,
     super.dropdownBuilder,
+    super.onBeforeClear,
+    super.onClear,
     //form properties
     super.onSaved,
     super.validator,
     super.decoratorProps,
+    super.selectedItemsWrapProps,
+    super.textProps,
   }) : super.multiSelection(
           popupProps: popupProps,
-          uiMode: UiMode.material,
+          uiMode: UiToApply.material,
           groupId: popupProps.autoCompleteProps.groupId,
         );
 }