+++ lib/src/properties/dropdown_props.dart
@@ -2,8 +2,7 @@ import 'package:dropdown_search/dropdown_search.dart';
 import 'package:dropdown_search/src/widgets/props/icon_button_props.dart';
 import 'package:flutter/material.dart';
 
-typedef DropdownButtonAnimationBuilder = Widget Function(
-    Widget child, bool isOpen);
+typedef DropdownButtonAnimationBuilder = Widget Function(Widget child, bool isOpen);
 
 Widget defaultAnimationBuilder(child, isOpen) {
   return AnimatedRotation(
@@ -42,6 +41,7 @@ class DropdownButtonProps extends IconButtonProps {
     super.style,
     super.isSelected,
     super.selectedIcon,
+    super.onHover,
   }) : super(icon: iconClosed);
 }