+++ lib/src/widgets/custom_cupertino_text_field.dart
@@ -79,6 +79,7 @@ class CustomCupertinoTextFields extends StatelessWidget {
       prefixMode: props.prefixMode,
       suffix: props.suffix,
       suffixMode: props.suffixMode,
+      selectAllOnFocus: props.selectAllOnFocus,
     );
   }
 }